backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
httpx==0.27.0
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
python-magic==0.4.27
